stuck because gameserver can't connect to redis
gameserver is started via gameserver-service and it isn't starting the container correctly. What's weird is that requests are routed to the container successfully, but it can't make requests outside the container to other microservices.

So either, it doesn't and just exposes it's state via an API. Downside of this is that it can't tell someone when it's "done", or whether it "failed".

We can also just ignore the problem for now, isolate all the redis state change information to it's own function that we can sort out later.

Will probably need to learn about how docker does networking.